show likes when logged out

+++ frontend/components/track/track.jsx
@@ -15,7 +15,6 @@ class Track extends React.Component {
     this.trackClickHandler = this.trackClickHandler.bind(this);
     this.likeClickHandler = this.likeClickHandler.bind(this);
     this.showIcon = this.showIcon.bind(this);
-    this.showHeart = this.showHeart.bind(this);
     this.playPauseIcon = this.playPauseIcon.bind(this);
     this.likeIcon = this.likeIcon.bind(this);
     this.commentForm = this.commentForm.bind(this);
@@ -60,11 +59,13 @@ class Track extends React.Component {
   }
 
   likeClickHandler () {
-    let currentLikes = this.state.likes;
-    if (this.props.track.likes.includes(this.props.currentUser.id)) {
-      this.props.deleteLike({track_id: this.props.track.id, user_id: this.props.currentUser.id});
-    } else {
-      this.props.createLike({track_id: this.props.track.id, user_id: this.props.currentUser.id});
+    if (this.props.currentUser) {
+      let currentLikes = this.state.likes;
+      if (this.props.track.likes.includes(this.props.currentUser.id)) {
+        this.props.deleteLike({track_id: this.props.track.id, user_id: this.props.currentUser.id});
+      } else {
+        this.props.createLike({track_id: this.props.track.id, user_id: this.props.currentUser.id});
+      }
     }
   }
 
@@ -78,14 +79,6 @@ class Track extends React.Component {
 
   showIcon () {
     if (this.trackUsername() === "You") {
-      return "show-icon";
-    } else {
-      return "hide-icon";
-    }
-  }
-
-  showHeart () {
-    if (this.props.currentUser) {
       return "show-icon";
     } else {
       return "hide-icon";
@@ -151,8 +144,8 @@ class Track extends React.Component {
             <div className='track-body-info-icons'>
               <img src="http://res.cloudinary.com/loudsounds/image/upload/c_scale,w_25/v1472928235/trash-512_tzepba.png" className={this.showIcon()} onClick={this.deleteClickHandler}></img>
               <img src="http://res.cloudinary.com/loudsounds/image/upload/c_scale,w_25/v1472928619/pencil-512_ddms2g.png" className={this.showIcon()} onClick={this.editClickHandler}></img>
-              <img src={this.likeIcon()} className={this.showHeart()} onClick={this.likeClickHandler}></img>
-              <span className={this.showIcon()}>{this.props.tracks[this.props.track.id].likes.length}</span>
+              <img src={this.likeIcon()} onClick={this.likeClickHandler}></img>
+              <span>{this.props.tracks[this.props.track.id].likes.length}</span>
             </div>
           </div>
         </div>
